chatbot/requirements.txt
absl-py==0.8.0
alembic==1.1.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
coverage==4.5.4
Flask==1.1.1
Flask-Injector==0.12.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
importlib-metadata==0.19
injector==0.17.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mecab-python3==0.996.2
more-itertools==7.2.0
numpy==1.17.2
opt-einsum==3.1.0
packaging==19.1
pluggy==0.12.0
protobuf==3.10.0
py==1.8.0
PyMySQL==0.9.3
pyparsing==2.4.2
pytest==5.1.1
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.8
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
typing==3.7.4.1
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
zipp==0.6.0
